start refactoring
- start banner
