+++ src/components/static/banner.js
@@ -1,23 +1,30 @@
 import React, { Component } from 'react';
+import FontAwesomeIcon from '@fortawesome/react-fontawesome';
+import faClock from '@fortawesome/fontawesome-free-solid/faClock';
+import faTv from '@fortawesome/fontawesome-free-solid/faTv';
+import faMoney from '@fortawesome/fontawesome-free-solid/faMoneyBillAlt';
+import faArrows from '@fortawesome/fontawesome-free-solid/faAngleDoubleDown';
+
+import '../../styles/static/banner.css';
 
 class banner extends Component {
 	render (){
 		return (
 			<div className="content--banner">
 				<div className="banner--fixed">
-					<div className="fixed--clear"></div>
+					<div className="fixed--clear"/>
 					<div className="fixed--title">
 						<p>Узнайте решение</p>
 						<span>и получите кредит</span>
 						<p>всего в 3 шага</p>
 						<div className="angle--down">
-							<i className="fas fa-angle-double-down"></i>
+                            <FontAwesomeIcon className='banner--arrows' icon={faArrows}/>
 						</div>
 					</div>
 					<div className="fixed--list">
 						<div className="list--application">
 							<div className="application--img">
-								<i className="fas fa-tv"></i>
+                                <FontAwesomeIcon className='banner--icons' icon={faTv}/>
 							</div>
 							<div className="application--text">
 								<p>Оставьте заявку</p>
@@ -25,7 +32,7 @@ class banner extends Component {
 						</div>
 						<div className="list--decision">
 							<div className="decision--img">
-								<i className="far fa-clock"></i>
+                                <FontAwesomeIcon className='banner--icons' icon={faClock}/>
 							</div>
 							<div className="decision--text">
 								<p>Узнайте решение</p>
@@ -33,7 +40,7 @@ class banner extends Component {
 						</div>
 						<div className="list--get">
 							<div className="get--img">
-								<i className="far fa-money-bill-alt"></i>
+                                <FontAwesomeIcon className='banner--icons' icon={faMoney}/>
 							</div>
 							<div className="get--text">
 								<p>Получите деньги в&nbsp;банке</p>
